+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## [2.23.2] - 2025-06-18
+
+-   Fix: EIH/EXT01/EXT02 cameras may display the original camera object after the camera is activated [0024618]
+-   Fix: Added a new Robot. When setting EIH in Activated Camera, there is a chance of error [0024619]
+-   Add: Case 6: Example of Prismatic Joint Gripper
+
 ## [2.23.1] - 2025-04-25
 
 -   Upgrade to support Isaac Sim 4.5.